refactor topic related classes, add mode
